Refactor: move stuff around, remove unused class
