+++ custom_components/kompas_energetyczny/config_flow.py
@@ -1,13 +1,15 @@
+"""Config Flow for Kompas Energetyczny"""
+
 import logging
 from homeassistant import config_entries
-from homeassistant.core import callback
-from homeassistant.data_entry_flow import FlowResultType
 import voluptuous as vol
-from .const import DOMAIN, DEFAULT_URL
+from .const import DOMAIN, API_URL_PRZESYLY
 
 _LOGGER = logging.getLogger(__name__)
 
 class KompasEnergetycznyConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
+    """Config Flow for KompasEnergetyczny"""
+
     VERSION = 1
 
     async def async_step_user(self, user_input=None):
@@ -17,7 +19,7 @@ class KompasEnergetycznyConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         if user_input is None:
             return self.async_show_form(
                 step_id="user",
-                data_schema=vol.Schema({vol.Optional("url", default=DEFAULT_URL): str}),
+                data_schema=vol.Schema({vol.Optional("url", default=API_URL_PRZESYLY): str}),
             )
 
         await self.async_set_unique_id(DOMAIN)
